+++ CHANGELOG.md
@@ -40,6 +40,8 @@ Compatibility:
 * Fix formatting in `Exception#full_message` when `RuntimeError` is not handled and `highlight` option is specified (@andrykonchin).
 * Fix `String#encode` and convert fallback values into String using `#to_str` (@andrykonchin).
 * Fix `Kernel.warn` and don't call `Warning#warn` if a specified category is disabled (@andrykonchin).
+* Fix `$!` global variable and make it fiber-local (@andrykonchin).
+* Fix `rb_set_errinfo` and `rb_errinfo` and store an error separately from `$!` (#2890, @andrykonchin).
 
 Performance: